Fixed so that Update only uses a single SaveChangesAsync call again (for atomicity).
Added missing checks for existing / non-existing saga data documents.

+++ Rebus.RavenDb/Sagas/RavenDbSagaStorage.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Security.Cryptography;
 using System.Text;
 using System.Threading.Tasks;
@@ -65,6 +66,14 @@ namespace Rebus.RavenDb.Sagas
             using (var session = _documentStore.OpenAsyncSession())
             {
                 var sagaDataDocumentId = SagaDataDocument.GetIdFromGuid(sagaData.Id);
+
+                var existingSagaDataDocument = await session.LoadAsync<SagaDataDocument>(sagaDataDocumentId);
+
+                if (existingSagaDataDocument != null)
+                {
+                    throw new ConcurrencyException("Cannot insert document with an id that already exists");
+                }
+
                 var sagaDataDocument = new SagaDataDocument(sagaData);
                 await session.StoreAsync(sagaDataDocument, sagaDataDocumentId);
 
@@ -77,19 +86,28 @@ namespace Rebus.RavenDb.Sagas
 
         public async Task Update(ISagaData sagaData, IEnumerable<ISagaCorrelationProperty> correlationProperties)
         {
-            SagaDataDocument existingSagaData;
             using (var session = _documentStore.OpenAsyncSession())
             {
                 var documentId = SagaDataDocument.GetIdFromGuid(sagaData.Id);
-                existingSagaData = await session.LoadAsync<SagaDataDocument>(documentId);
+                var existingSagaData = await session.LoadAsync<SagaDataDocument>(documentId);
+
+                if (existingSagaData == null)
+                {
+                    throw new ConcurrencyException("Cannot update saga that does not exist");
+                }
+
                 sagaData.Revision++;
                 existingSagaData.SagaData = sagaData;
-
-                await DeleteCorrelationPropertyDataForSaga(existingSagaData, session);
-                await session.SaveChangesAsync();
             
                 //add the new saga correlation documents
                 var correlationPropertyDocumentIds = await SaveCorrelationProperties(session, sagaData, correlationProperties, existingSagaData.Id);
+
+                var oldCorrelationPropertyDocumentIdsNotPresentInNew =
+                    existingSagaData.SagaCorrelationPropertyDocumentIds.Where(
+                        sc => !correlationPropertyDocumentIds.Contains(sc));
+
+                await DeleteCorrelationProperties(oldCorrelationPropertyDocumentIdsNotPresentInNew, session);
+
                 existingSagaData.SagaCorrelationPropertyDocumentIds = correlationPropertyDocumentIds;
 
                 await session.SaveChangesAsync();
@@ -102,6 +120,12 @@ namespace Rebus.RavenDb.Sagas
             {
                 var documentId = SagaDataDocument.GetIdFromGuid(sagaData.Id);
                 var existingSagaData = await session.LoadAsync<SagaDataDocument>(documentId);
+
+                if (existingSagaData == null)
+                {
+                    throw new ConcurrencyException("Cannot delete saga that does not exist");
+                }
+
                 await DeleteCorrelationPropertyDataForSaga(existingSagaData, session);
 
                 session.Delete(existingSagaData);
@@ -109,13 +133,19 @@ namespace Rebus.RavenDb.Sagas
                 await session.SaveChangesAsync();
             }
         }
-
+        
         private async Task DeleteCorrelationPropertyDataForSaga(SagaDataDocument sagaDataDocument,
+            IAsyncDocumentSession session)
+        {
+            await DeleteCorrelationProperties(sagaDataDocument.SagaCorrelationPropertyDocumentIds, session);
+        }
+
+        private async Task DeleteCorrelationProperties(IEnumerable<string> correlationPropertyIds,
             IAsyncDocumentSession session)
         {
             var existingSagaCorrelationPropertyDocuments =
                 await session.LoadAsync<SagaCorrelationPropertyDocument>(
-                    sagaDataDocument.SagaCorrelationPropertyDocumentIds);
+                    correlationPropertyIds);
 
             //delete the existing saga correlation documents
             foreach (var existingSagaCorrelationPropertyDocument in existingSagaCorrelationPropertyDocuments)
@@ -141,15 +171,21 @@ namespace Rebus.RavenDb.Sagas
                     await session.LoadAsync<SagaCorrelationPropertyDocument>(documentId);
 
                 if (existingSagaCorrelationPropertyDocument != null)
-                    throw new ConcurrencyException(
-                        $"Could not save correlation properties. The following correlation property already exists with the same value for another saga: {propertyName} - {value}");
+                {
+                    if (existingSagaCorrelationPropertyDocument.SagaDataDocumentId != sagaDataDocumentId)
+                    {
+                        throw new ConcurrencyException(
+                            $"Could not save correlation properties. The following correlation property already exists with the same value for another saga: {propertyName} - {value}");
+                    }
+                }
+                else
+                {
+                    var sagaCorrelationPropertyDocument = new SagaCorrelationPropertyDocument(correlationProperty.SagaDataType, propertyName,
+                        value, sagaDataDocumentId);
 
-                var sagaCorrelationPropertyDocument = new SagaCorrelationPropertyDocument(correlationProperty.SagaDataType, propertyName,
-                    value, sagaDataDocumentId);
-
-                await session.StoreAsync(sagaCorrelationPropertyDocument, documentId);
-
-
+                    await session.StoreAsync(sagaCorrelationPropertyDocument, documentId);
+                }
+                
                 documentIds.Add(documentId);
             }
